requirements.txt
django
djangorestframework
django-allauth
dj-rest-auth
environs
pillow
python-dotenv
requests
cryptography
social-auth-app-django
pipenv
django-cors-headers
python-decouple
psycopg2-binary
djangorestframework-simplejwt
google-auth
google-auth-oauthlib
google-auth-httplib2
gunicorn
whitenoise